requirements.txt
benchtemp==1.1.1
dgl==2.1.0
matplotlib==3.8.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
Pillow==10.3.0
PyYAML==6.0.1
ruamel.base==1.0.0
scikit_learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==68.0.0
tensorboard==2.15.1
torch==2.3.0
torch_geometric==2.5.3
tqdm==4.65.0
